rename api project to reviewer_api
