+++ api/README.md
@@ -1,3 +1,3 @@
-# Request Management API
+# FOI DOC REVIEWER API
 
 README todo
