+++ api/manage.py
@@ -23,7 +23,7 @@ from flask_script import Manager  # class for handling a set of commands
 # models included so that migrate can build the database migrations
 from request_api import models  # pylint: disable=unused-import
 from request_api import create_app
-from request_api.models import db
+from reviewer_api.models import db
 
 ## python manage.py db migrate
 ## python manage.py db upgrade
