+++ api/reviewer_api/auth.py
@@ -18,7 +18,7 @@ from http import HTTPStatus
 from flask import g, request
 from flask_jwt_oidc import JwtManager
 from jose import jwt as josejwt
-from request_api.utils.enums import MinistryTeamWithKeycloackGroup, ProcessingTeamWithKeycloackGroup, IAOTeamWithKeycloackGroup
+from reviewer_api.utils.enums import MinistryTeamWithKeycloackGroup, ProcessingTeamWithKeycloackGroup, IAOTeamWithKeycloackGroup
 jwt = (
     JwtManager()
 )  # pylint: disable=invalid-name; lower case name as used by convention in most Flask apps
