+++ api/reviewer_api/schemas/__init__.py
@@ -11,25 +11,5 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-"""Meta information about the service.
+"""Schema package."""
 
-Currently this only provides API versioning information
-"""
-from flask import jsonify
-from flask_restx import Namespace, Resource
-
-from request_api.utils.run_version import get_run_version
-
-
-API = Namespace('Meta', description='Metadata')
-
-
-@API.route('/info')
-class Info(Resource):
-    """Meta information about the overall service."""
-
-    @staticmethod
-    def get():
-        """Return a JSON object with meta information about the Service."""
-        version = get_run_version()
-        return jsonify(API=f'request_api/{version}')
