+++ api/reviewer_api/utils/run_version.py
@@ -1,0 +1,29 @@
+# Copyright © 2021 Province of British Columbia
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+"""Supply version and commit hash info."""
+import os
+
+from reviewer_api.version import __version__
+
+
+def _get_build_openshift_commit_hash():
+    return os.getenv('OPENSHIFT_BUILD_COMMIT', None)
+
+
+def get_run_version():
+    """Return a formatted version string for this service."""
+    commit_hash = _get_build_openshift_commit_hash()
+    if commit_hash:
+        return f'{__version__}-{commit_hash}'
+    return __version__
